Don't pipe to sort, run it from find.
If the sort will fail, for example (the real life example) on ' /sys/devices/platform/Fixed: no such file', the find will be killed. For some reason the file have space in name, or something. Using -exec fixes it. Thanks for Oleg Vinichenko for all the testing needed, on openvz kernel.